ARRANGE and MOVE BioSkills.js and Projects.js into a new folder
